requirements.txt
Flask
Flask-Compress
Gunicorn
Jinja2
pandas
gspread
oauth2client
python-dotenv
Pillow